+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   existing big-endian module.
 * Add functions for parsing arrays of big/little endian words of
   various sizes.
+* Add `skipUntil` to `Latin`.
 
 ## 0.3.0.0 -- 2019-09-30